Added JDWP step over stress test
This test tries to step over a method call that iterates over every
annotation of a class with 1000 methods.

On some VM implementations (e.g. ART) the annotation processing code
is implemented in Java, and if all the code is executed in interpreted
mode during a step over, this test may run so long that it makes the
debugging process unusable.

Change-Id: Id4daf69edc7f0ce7bd9ca7ce6cec6fb48758ec03
Signed-off-by: Daniel Mihalyi <daniel.mihalyi@mattakis.com>