Update makefile for JDWP tests
This CL updates the makefile so we can run JDWP tests on each ABI
supported by the host and the target. It adds ABI-specific rules:
* run-jdwp-tests-host32 and run-jdwp-tests-target32 for 32-bit ABI
* run-jdwp-tests-host64 and run-jdwp-tests-target64 for 64-bit ABI
The existing run-jdwp-tests-host and run-jdwp-tests-target now run
tests for each ABI.

It also updates AllTests suite to System.exit(1) in case of failure
or error so the make fails. This allows to catch regression during
continuous testing on the host.

Bug: 18406998
Change-Id: I9b4273783ac01e59d3c50242fcaf22ea8b325300