Update makefile for JDWP tests
This CL updates the makefile so we can run JDWP tests on each ABI
supported by the host and the target.
It adds the following ABI-specific rules for host and target:
* run-jdwp-tests-host32 and run-jdwp-tests-target32 for 32-bit ABI
* run-jdwp-tests-host64 and run-jdwp-tests-target64 for 64-bit ABI

The existing run-jdwp-tests-host and run-jdwp-tests-target now run
tests for each ABI. To avoid interleaving the output, we need to
run them sequentally. Therefore, we create variants of the new rules
above so they depend on the previous ABI. Here is an example for host
supporting 64-bit and 32-bit ABIs:
run-jdwp-tests-host
 -> run-jdwp-tests-host-all32
   -> run-jdwp-tests-host-all64
This will execute 64-bit host tests before 32-bit host tests.

It also updates AllTests suite to System.exit(1) in case of failure
or error so the make fails. This allows to catch regression during
continuous testing on the host.

Bug: 18406998
Change-Id: I9b4273783ac01e59d3c50242fcaf22ea8b325300