Fix ThreadOnlyModifierTest flakiness
Tests in ThreadOnlyModifierTest class sometimes fail with a warning
message that the connection with the debuggee is closed. This means
we could not send the final VirtualMachine.Dispose command to the
debuggee. That may happen if the debuggee process has already exited
after we resume it during the test.

To avoid that situation, we no longer resume the debuggee (with a
VirtualMachine.Resume command) at the end of the test.

Also computes class name and signature at runtime instead of having
them hardcoded in constant strings. This is helpful for refactoring.

Bug: 19628620
Change-Id: Id68d7f57521b12d11aa37605f1282bf39a5ab202