[WIP] JDWP: only show output on failure on the host
Redirects test output to a temp file and only prints the content of
this file on failure when running tests with run-jdwp-tests-host.
Also lists passed and failed tests so it is easier to control in a
continuous build context.

For ABi-specific target (like run-jdwp-tests-host64), this will also
print the output of passed tests. This is useful when adding extra
tests.

The support for the target tests (on device) will be added later.

Bug: 18406998
Change-Id: I76abffdcecd89db91ce1f9c8e368866183222c5c