JDWP: support continuous host testing
Prints output of JDWP tests only when there is a failure. It
creates a temp directory on the host and redirects test output to
a temp file.

To prepare for continuous testing, we list passed and failed tests
at the end so we can quickly catch failures. It is close to testing
done in ART with test-art-host. However, since we run a whole JUnit
test suite, we do not list individual tests as passed or failed. We
can only say whether all tests passed for the tested ABI.

For development purpose (like adding extra tests), the ABI-specific
targets (like run-jdwp-tests-host64) prints output even if tests
passed. It is helpful to verify the sequence of commands during a
new test.

The support for on-device testing will be added later (we already
have CTS for that).

Bug: 18406998
Change-Id: I76abffdcecd89db91ce1f9c8e368866183222c5c