JDWP: test invocation with thread suspension
Tests the debuggee can still process new commands while a thread is
blocked invoking a command (because another event in another thread
suspended that thread). The tested commands are:
- ClassType.InvokeCommand
- ClassType.NewInstance
- ObjectReference.InvokeCommand

Also tests the case where the debugger detaches while a thread is
invoking a method on its behalf (with one of the command above). The
method invocation completes but no reply is sent and the thread does
not suspend itself or other threads now the debugger is no longer
attached.

Bug: 21515842
Change-Id: Ifd680fdc4b0dfd72d449cce692b4229eca02cdce