[DO NOT SUBMIT] JDWP: test setting constant local variables
Adds test org.apache.harmony.jpda.tests.jdwp.StackFrame.SetValues002Test#testSetValues005_IntConstant
that will fail with Quick and Optimizing compilers because we change
the value of a local variable seen as a constant.

Also changes the way we suspend the debuggee. We now synchronize
through a socket to suspend the debuggee instead of suspending it
on a breakpoint.

Bug: 19944235
Change-Id: I88432133aec4225f5ce079086b60c49a209043c7