JDWP: fail test if debuggee does not exit normally
Makes a test fail if the debuggee exits with an error. This prevents
from passing a test while the debuggee aborts (during runtime
shutdown for instance).

By default, we expect the debuggee to exit with code 0. For tests
realted to VirtualMachine.Exit command (that forces the runtime to
exit with a given error code), we add a way to specify which is the
expected exit code.

Bug: 22907762
Bug: 17941266
Change-Id: I1acd4a4251d8cb3a335e581aef1792ea1049ba74