JDWP Tests: More GetClassLoader tests
First, ensure that the returned object (in the success case) is
a classloader.

Second, add a negative test that ensures we get the right failure
value when the supplied ID is not a reference type. Note: this only
works if reference type IDs and object type IDs share their domain.
Otherwise a valid object ID may be a valid reference type ID.

Bug: 26349019
Change-Id: Ia33dac3f8b8fd158dec9f229f796263c8c985eab