Enable Apache Harmony crypto tests.
A couple tests needed to be modified to handle the fact that we don't use
Apache Harmony as our implementation.  As well, we jarjar our classes so
serialization tests needed to have their golden files moved.

These tests rely on classes in security/support, hence that being added
to the build as well.

Bug: 4483710
Test: cts -m CtsLibcoreTestCases
Change-Id: I2ffad403a495a7a0fafbf968ffae6094d43966f7