JDWP: allow to set debuggee classpath
Adds property 'jpda.settings.debuggeeClasspath' to override the
classpath set for the debuggee (by default, this is the classpath
of the process starting the debuggee).

This is a convenient way to run the junit tests with java runtime
(in an IDE for instance) while the debuggee still runs on ART
runtime.

Bug: 36712802
Test: manual
Change-Id: I30c94bf8c8fed7ddc8adb9f04b528522268eb5ff