JDWP: update ReferenceType.InstancesTest
Updates debuggee of test ReferenceType.InstancesTest to make sure
unreachable objects cannot be kept alive by a DEX register in the
stack frame used by the test.

Indeed the last 'new' in a loop can be stored in a DEX register which is
not overwritten after that instruction. To avoid that situation, we move
the allocation of such objects into a separate method.

Bug: 36921933
Test: art/tools/run-jdwp-tests.sh --mode=host --variant=X64
Change-Id: I723a83430121a02470fa8bdcc9af77ec358d4929