JDWP: prioritize usage of RT.MethodsWithGeneric command
The methods of a class are accessible with the commands
ReferenceType.Methods and ReferenceType.MethodsWithGeneric. However
older runtimes (like Dalvik) only supports the generic version. Since
some helper methods of the jdwp test framework only use the
non-generic command, we cannot run the jdwp tests on older runtimes.

This CL fixes that by changing the helper methods to use the command
MethodsWithGeneric first. If it's not supported by the runtime, we
fallback to the command Methods.

Note that the tests that specifically use one or the other command
are unchanged.

Bug: 65869324
Test: art/tools/run-jdwp-tests.sh '--mode=host' '--variant=X64'
Change-Id: I08b50f79385fab5071ea9b7d6cfb80744d059779