JDWP: prioritize usage of Method.VariableTableWithGeneric command
The variables in a method are accessible with the commands
Method.VariableTable and Method.VariableTableWithGeneric. However
older runtimes (like Dalvik) only supports the generic version. Since
some helper methods of the jdwp test framework only use the
non-generic command, we cannot run the jdwp tests on older runtimes.

This CL fixes that by changing the helper methods to use the command
VariableTableWithGeneric first. If it is not supported by the
runtime, we fallback to the command VariableTable.

Also removes the VarInfo class which is redundant with the class
Frame.Variable.

Note that the tests that specifically use one or the other command
are unchanged.

Bug: 65869324
Test: art/tools/run-jdwp-tests.sh '--mode=host' '--variant=X64'
Change-Id: If69ac488c6a6ffd0d3f5ce76e9b9415a8d3b62ca