Add test for DDM.Chunk jdwp extension command
This is a long-standing extension in dalvik and art. It allows one to
interact with the DdmServer system. The command-set is -57 (119u) and
the command is (1). It has a variable return type with either no
information or a ddms packet.

This test was added to prevent regressions since this command is very
strange.

Test: ./tools/art/run-jdwp-tests.sh --mode=host
Bug: 69169846
Change-Id: I86252ba9859bb28ab5ebabb730fa8efad2e8ca63