Remove overly strict slot check in JDWPMethodVariabeTableTestCase
The test was checking that all method arguments would appear in the
first slots of the method variable table. This is a common arrangement
but it is not required by the JDWP specification or the JLS. Some
runtimes, such as ART, organize them differently.  Testing shows that
real-world JDWP debuggers do not rely on the initial ordering of
arguments.

Bug: 66905468
Test: ./art/tools/run-libjdwp-tests.sh --mode=host
Test: ./art/tools/run-jdwp-tests.sh --mode=host
Change-Id: Ica1a351d98fc614f21dcf7995ef8d84b2401dd75