Allow jdwp tests to dump native traces on timeout.
Often timeouts are pretty much impossible to debug because we don't
get any real information about what the debuggee is doing. To help
with debugging issues like this we give the test runner the ability to
try to dump a stack using a user supplied program.

Users should set the jpda.settings.dumpProcess property to a program
capable of dumping a pid or forcing the pid to dump its own stacks.

On android this is either 'kill -3' or 'su root debuggerd -b'.

Test: modify a test with infinite loop, Run
Bug: 69243589
Bug: 70838465
Bug: 72928430
Change-Id: I54dfddf29678107b2480a78587a66ebdd87a7447